Added original SqueezeCenter license
Added SqueezeCenter license file since this project include some
verbatim and modified versions of source files included in SqueezeCenter
(most notably the patch_src directory, used for building the
Gentoo-specific patches).

This is mainly because this project is pushed to Github and hence is
effectively redistributing those files.

The files in question are all licensed under the GPL-2.