Switch back to mplayer (bug#277025)
Gentoo has to switch back to mplayer because the version of 'faad' in
Portage does not include the patches applied to the bundled copy that
comes with SqueezeCenter. Since Gentoo can't use the bundled copy this
gives problems (Gentoo bug#277025).

We'll hopefully switch back at some point when those FAAD changes are
incorporated upstream.